requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.3
Django==5.2.1
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-mdeditor==0.1.20
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.16.0
idna==3.10
pillow==11.2.1
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
requests==2.32.3
setuptools==80.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
whitenoise==6.9.0
wsproto==1.2.0
    